LPS-98855 Extract search subcomponents
